JSON: Add a / at the end of base name.

+++ core/utils.c
@@ -219,7 +219,7 @@ size_t utils_floatToText(double data,
     int64_t intPart;
     double decPart;
 
-    if (data <= INT64_MIN || data >= INT64_MAX) return 0;
+    if (data <= (double)INT64_MIN || data >= (double)INT64_MAX) return 0;
 
     intPart = (int64_t)data;
     decPart = data - intPart;
